start moving pagination logic out of svelte core
