+++ packages/houdini-react/src/runtime/hooks/usePaginationHandlers.ts
@@ -1,0 +1,7 @@
+import { QueryArtifact } from '$houdini/runtime/lib/types'
+
+import { useDocumentStore } from './useDocumentStore'
+
+export function usePaginationHandlers({ artifact }: { artifact: QueryArtifact }) {
+	const [storeValue, observer] = useDocumentStore({ artifact })
+}
