+++ packages/houdini/src/runtime/lib/pagination.ts
@@ -1,23 +1,18 @@
-import type { SendParams } from '$houdini/runtime/client/documentStore'
-import { CachePolicy } from '$houdini/runtime/lib'
-import { getCurrentConfig } from '$houdini/runtime/lib/config'
-import { siteURL } from '$houdini/runtime/lib/constants'
-import { deepEquals } from '$houdini/runtime/lib/deepEquals'
-import type { GraphQLObject, QueryArtifact, QueryResult } from '$houdini/runtime/lib/types'
-import { writable } from 'svelte/store'
-
+import { CachePolicy } from '.'
 import { getSession } from '../../session'
-import type { CursorHandlers } from '../../types'
+import type { SendParams } from '../client/documentStore'
+import type { GraphQLObject, QueryArtifact, QueryResult } from '../lib/types'
 import type { QueryStoreFetchParams } from '../query'
 import { fetchParams } from '../query'
-import type { FetchFn } from './fetch'
-import type { PageInfo } from './pageInfo'
+import { getCurrentConfig } from './config'
+import { siteURL } from './constants'
+import { deepEquals } from './deepEquals'
 import { countPage, extractPageInfo, missingPageSizeError } from './pageInfo'
+import type { CursorHandlers, FetchFn } from './types'
 
 export function cursorHandlers<_Data extends GraphQLObject, _Input extends Record<string, any>>({
 	artifact,
 	storeName,
-	initialValue,
 	fetchUpdate: parentFetchUpdate,
 	fetch: parentFetch,
 	getState,
@@ -28,11 +23,8 @@ export function cursorHandlers<_Data extends GraphQLObject, _Input extends Recor
 	fetch: FetchFn<_Data, _Input>
 	getState: () => _Data | null
 	getVariables: () => _Input
-	initialValue: _Data | null
 	fetchUpdate: (arg: SendParams, updates: string[]) => ReturnType<FetchFn<_Data, _Input>>
-}): CursorHandlers<_Data, _Input> {
-	const pageInfo = writable<PageInfo>(extractPageInfo(initialValue, artifact.refetch!.path))
-
+}): Omit<CursorHandlers<_Data, _Input>, 'pageInfo'> {
 	// dry up the page-loading logic
 	const loadPage = async ({
 		pageSizeVar,
@@ -93,9 +85,6 @@ export function cursorHandlers<_Data extends GraphQLObject, _Input extends Recor
 			// make sure that we pull the value out of the correct query field
 			resultPath.unshift(config.types[targetType].resolve!.queryField)
 		}
-
-		// we need to find the connection object holding the current page info
-		pageInfo.set(extractPageInfo(data, resultPath))
 	}
 
 	const getPageInfo = () => {
@@ -188,7 +177,6 @@ If you think this is an error, please open an issue on GitHub`)
 				where: 'start',
 			})
 		},
-		pageInfo,
 		async fetch(
 			args?: QueryStoreFetchParams<_Data, _Input>
 		): Promise<QueryResult<_Data, _Input>> {
@@ -263,10 +251,107 @@ Make sure to pass a cursor value by hand that includes the current set (ie the e
 				variables: queryVariables as _Input,
 			})
 
-			// keep the page info store up to date
-			pageInfo.set(extractPageInfo(result.data, artifact.refetch!.path))
-
 			return result
+		},
+	}
+}
+
+export function offsetHandlers<_Data extends GraphQLObject, _Input extends {}>({
+	artifact,
+	storeName,
+	getState,
+	getVariables,
+	fetch: parentFetch,
+	fetchUpdate: parentFetchUpdate,
+}: {
+	artifact: QueryArtifact
+	fetch: FetchFn<_Data, _Input>
+	fetchUpdate: (arg: SendParams) => ReturnType<FetchFn<_Data, _Input>>
+	storeName: string
+	getState: () => _Data | null
+	getVariables: () => _Input
+}) {
+	// we need to track the most recent offset for this handler
+	let getOffset = () =>
+		(artifact.refetch?.start as number) ||
+		countPage(artifact.refetch!.path, getState()) ||
+		artifact.refetch!.pageSize
+
+	let currentOffset = getOffset() ?? 0
+
+	return {
+		loadNextPage: async ({
+			limit,
+			offset,
+			fetch,
+			metadata,
+		}: {
+			limit?: number
+			offset?: number
+			fetch?: typeof globalThis.fetch
+			metadata?: {}
+		} = {}) => {
+			// build up the variables to pass to the query
+			const queryVariables: Record<string, any> = {
+				...getVariables(),
+				offset: offset ?? getOffset(),
+			}
+			if (limit || limit === 0) {
+				queryVariables.limit = limit
+			}
+
+			// if we made it this far without a limit argument and there's no default page size,
+			// they made a mistake
+			if (!queryVariables.limit && !artifact.refetch!.pageSize) {
+				throw missingPageSizeError('loadNextPage')
+			}
+
+			// Get the Pagination Mode
+			let isSinglePage = artifact.refetch?.mode === 'SinglePage'
+
+			// send the query
+			const targetFetch = isSinglePage ? parentFetch : parentFetchUpdate
+			await targetFetch({
+				variables: queryVariables as _Input,
+				fetch,
+				metadata,
+				policy: isSinglePage ? artifact.policy : CachePolicy.NetworkOnly,
+				session: await getSession(),
+			})
+
+			// add the page size to the offset so we load the next page next time
+			const pageSize = queryVariables.limit || artifact.refetch!.pageSize
+			currentOffset = offset + pageSize
+		},
+		async fetch(
+			args?: QueryStoreFetchParams<_Data, _Input>
+		): Promise<QueryResult<_Data, _Input>> {
+			const { params } = await fetchParams(artifact, storeName, args)
+
+			const { variables } = params ?? {}
+
+			// if the input is different than the query variables then we just do everything like normal
+			if (variables && !deepEquals(getVariables(), variables)) {
+				return parentFetch.call(this, params)
+			}
+
+			// we are updating the current set of items, count the number of items that currently exist
+			// and ask for the full data set
+			const count = currentOffset || getOffset()
+
+			// build up the variables to pass to the query
+			const queryVariables: Record<string, any> = {}
+
+			// if there are more records than the first page, we need fetch to load everything
+			if (!artifact.refetch!.pageSize || count > artifact.refetch!.pageSize) {
+				queryVariables.limit = count
+			}
+
+			// send the query
+			return await parentFetch.call(this, {
+				...params,
+				variables: queryVariables as _Input,
+			})
 		},
 	}
 }
